requirements.txt
aiogram==3.0.0
ytmusicapi
yt-dlp
python-dotenv
ffmpeg-python